updated branching capabilities
