+++ goblet_workflows/__init__.py
@@ -1,1 +1,3 @@
 from goblet_workflows.workflow import Workflow  # noqa: F401
+from goblet_workflows.steps import Step, MultiStep, AssignStep, HttpStep, BQStep # noqa: F401
+from goblet_workflows.controls import Branch, For # noqa: F401
