+++ goblet_workflows/__version__.py
@@ -1,4 +1,4 @@
-VERSION = (0, 1, 1)
+VERSION = (0, 1, 2)
 
 
 __version__ = ".".join(map(str, VERSION))
